# backend/swagger.yaml
swagger: "2.0"
info:
  description: "Trainer REST API documentation"
  version: "1.0.0"
  title: "Trainer"
host: "trainer.ksi.fit.cvut.cz"
basePath: "/api"
...
schemes:
- "https"
paths:
...
  /weeks:
    post:
      tags:
        - "weeks"
      summary: "Create new week"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "Week to be created"
          required: true
          schema:
            $ref: "#/definitions/WeekCreateDTO"
      responses:
        "201":
          description: "Success, week created"
          schema:
            $ref: "#/definitions/WeekGetDTO"
        "401":
          description: "Unauthorized user"
        "403":
          description: "Insufficient rights to create new week (not a teacher)"
      security:
        - login_secret: [ ]
...
  /weeks/{id}/lessons:
    patch:
      tags:
        - "weeks"
      summary: "Edit existing week lesson order"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "path"
          description: "Unique week identifier"
          name: "id"
          type: "integer"
          required: true
        - in: "body"
          name: "body"
          description: "Lesson ids to be put into week with order"
          required: true
          schema:
            $ref: "#/definitions/WeekLessonOrderDTO"
      responses:
        "204":
          description: "Success, lessons edited"
        "400":
          description: "Tried to update lessons outside week's course"
        "403":
          description: "Insufficient rights to edit the week"
        "404":
          description: "Week or lesson with given identifier not found"
      security:
        - login_secret: [ ]
  /weeks/{id}/courses/{courseId}:
    post:
      tags:
        - "weeks"
      summary: "Clone week into new course"
      produces:
        - "application/json"
      parameters:
        - in: "path"
          description: "Unique week identifier"
          name: "id"
          type: "integer"
          required: true
        - in: "path"
          description: "Unique course identifier"
          name: "courseId"
          type: "integer"
          required: true
      responses:
        "200":
          description: "Success, week cloned"
          schema:
            $ref: "#/definitions/WeekGetDTO"
        "403":
          description: "Insufficient rights to edit the week or add week to the course"
        "404":
          description: "Week or course with given identifier not found"
      security:
        - login_secret: [ ]

  # Lessons
...
  /code:
    post:
      tags:
        - "code-module"
      summary: "Create new code module"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "Code module to be created"
          required: true
          schema:
            $ref: "#/definitions/CodeModuleCreateDTO"
      responses:
        "201":
          description: "Success, code module created"
          schema:
            $ref: "#/definitions/CodeModuleReadDTO"
        "401":
          description: "Unauthorized user"
        "403":
          description: "Insufficient rights to create new code module"
        "404":
          description: "Matching module not found"
      security:
        - login_secret: [ ]
...
  /code/tests/{testId}:
    delete:
      tags:
        - "code-module"
      summary: "Delete existing code module test"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "path"
          description: "Unique code module test identifier"
          name: "testId"
          type: "integer"
          required: true
      responses:
        "204":
          description: "Success, code module test deleted"
        "401":
          description: "Unauthorized user"
        "403":
          description: "Insufficient rights to delete the code module test"
        "404":
          description: "Code module test with given identifier not found"
      security:
        - login_secret: [ ]
  /code/files/{fileId}:
    delete:
      tags:
        - "code-module"
      summary: "Delete existing code module file"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "path"
          description: "Unique code module file identifier"
          name: "fileId"
          type: "integer"
          required: true
      responses:
        "204":
          description: "Success, code module file deleted"
        "401":
          description: "Unauthorized user"
        "403":
          description: "Insufficient rights to delete the code module file"
        "404":
          description: "Code module file with given identifier not found"
      security:
        - login_secret: [ ]
securityDefinitions:
  login_secret:
    type: "apiKey"
    name: "LoginSecret"
    in: "header"
definitions:
...
  CourseFindDTO:
    description: "Course list DTO"
    type: "object"
    properties:
      id:
        type: "integer"
        description: "Unique course identifier"
        example: 1
      name:
        type: "string"
        description: "Course name"
        example: "Programming and Algorithms 1 (3 + 4)"
      subject:
        $ref: "#/definitions/SubjectFindDTO"
      semester:
        $ref: "#/definitions/SemesterFindDTO"
      lessonCompleted:
        type: "integer"
        description: "Number of completed lessons in course"
        example: 2
      lessonCount:
        type: "integer"
        description: "Number of lessons in course"
        example: 5
      role:
        $ref: "#/definitions/RoleEnum"
...
  SubjectFindDTO:
    description: "Subject DTO"
    type: "object"
    properties:
      id:
        type: "integer"
        description: "Unique subject identifier"
        example: 1
      name:
        type: "string"
        description: "Subject name"
        example: "Programming and Algorithms 1"
      code:
        type: "string"
        description: "Subject code"
        example: "BI-PA1.21"
  SemesterFindDTO:
    description: "Semester DTO"
    type: "object"
    properties:
      id:
        type: "integer"
        description: "Unique semester identifier"
        example: 1
      code:
        type: "string"
        description: "Semester code"
        example: "B232"
      from:
        type: "string"
        description: "Semester start date"
        format: "date-time"
      until:
        type: "string"
        description: "Semester end date"
        format: "date-time"

  # Week
  WeekFindDTO:
    description: "Week"
    type: "object"
    properties:
      id:
        type: "integer"
        description: "Unique week identifier"
        example: 1
      name:
        type: "string"
        description: "Week name"
        example: "1st week"
      from:
        type: "string"
        format: "date-time"
        description: "Week start date"
        example: "2024-02-01T00:00:00.000+00:00"
      until:
        type: "string"
        format: "date-time"
        description: "Week end date"
        example: "2025-02-01T00:00:00.000+00:00"
      course:
        $ref: "#/definitions/CourseFindDTO"
...
  WeekGetDTO:
    description: "Week"
    allOf:
      - $ref: "#/definitions/WeekFindDTO"
    properties:
      lessons:
        $ref: "#/definitions/LessonFindDTOArray"
...
  WeekUpdateDTO:
    description: "Week to be updated"
    type: "object"
    properties:
      name:
        type: "string"
        description: "Week name"
        example: "First week"
      from:
        type: "string"
        format: "date-time"
        description: "Week start date"
        example: "2024-01-01T00:00:00.000+00:00"
      until:
        type: "string"
        format: "date-time"
        description: "Week end date"
        example: "2025-01-01T00:00:00.000+00:00"
  WeekCreateDTO:
    description: "Week to be created"
    allOf:
      - $ref: "#/definitions/WeekUpdateDTO"
    properties:
      courseId:
        type: "integer"
        description: "Unique identifier of course to put the new week into"
        example: 2
  WeekLessonOrderDTO:
    type: "object"
    description: "List of lessons to be updated"
    properties:
      lessonIds:
        description: "Array of lessons to be updated (order matters)"
        type: "array"
        items:
          type: "integer"
        example: [4, 2, 1, 3]
  # Course user
...
  LessonDTO:
    type: "object"
    properties:
      id:
        type: "integer"
        description: "Unique lesson identifier"
        example: 1
      name:
        type: "string"
        description: "Lesson name"
        example: "Compilation"
      hidden:
        type: "boolean"
        description: "Is lesson hidden from students?"
        example: false
      type:
        $ref: "#/definitions/LessonType"
      lockCode:
        type: "string"
        description: "Code for unlocking modules in lesson (teacher only)"
        example: null
      timeLimit:
        type: "string"
        format: "date-time"
        description: "Lesson end date"
        example: "2025-02-01T00:00:00.000+00:00"
  LessonFindDTOArray:
    type: "array"
    items:
      $ref: "#/definitions/LessonFindDTO"
  LessonFindDTO:
    description: "Lesson"
    type: "object"
    allOf:
      - $ref: "#/definitions/LessonDTO"
    properties:
      order:
        type: "integer"
        description: "Lesson order in week"
        example: 1
      progress:
        type: "integer"
        description: "Overall progress in lesson"
        example: 50
...
  LessonType:
    type: "string"
    enum: [ "TUTORIAL", "TUTORIAL_PREPARATION", "INDIVIDUAL_TASK", "SUPPLEMENTARY", "FREQUENT_MISTAKES", "INFORMATION" ]
    description: "Lesson type"
    example: "TUTORIAL"
...
  RoleEnum:
    type: "string"
    description: "User role in course"
    enum: ["TEACHER", "STUDENT"]
    example: "STUDENT"

  # User
...
  CodeModuleDTO:
    type: "object"
    properties:
      codeType:
        type: "string"
        description: "Type of code module"
        enum: ["SHOWCASE", "TEST_IO", "TEST_ASSERT", "WRITE_ASSERT"]
        example: "TEST_ASSERT"
      interactionType:
        type: "string"
        description: "Possible types of interaction with the module"
        enum: ["UPDOWN_ONLY", "EDITOR"]
        example: "EDITOR"
      codeHidden:
        type: "string"
        description: "Code used to test student code"
        example: "#include <stdio.h> int main () { return 0; }"
      referencePublic:
        type: "boolean"
        description: "Parameter defining whether reference solution should be downloadable"
        example: false
      fileLimit:
        type: "integer"
        description: "Maximum file size for file uploading"
        example: 10240
      libraryType:
        type: "string"
        description: "Type of library linked to resulting code"
        enum: ["NO_LIB", "LIB_C", "LIB_CPP"]
        example: "LIB_C"
      envelopeType:
        type: "string"
        enum: ["ENV_C", "ENV_C_IO", "ENV_CPP", "ENV_CPP_STL", "ENV_CUSTOM"]
        description: "Envelope into which student/ref codes are inserted"
        example: "ENV_C_IO"
      customEnvelope:
        type: "string"
        description: "If custom envelope is used, its content is here"
        example: null
  CodeModuleReadDTO:
    description: "Code module information DTO"
    allOf:
      - $ref: "#/definitions/CodeModuleDTO"
    properties:
      id:
        type: "integer"
        description: "Unique code module identifier"
        example: 1
      tests:
        type: "array"
        items:
          $ref: "#/definitions/CodeModuleTestDTO"
      files:
        type: "array"
        items:
          $ref: "#/definitions/CodeModuleFileDTO"
  CodeModuleUpdateDTO:
    description: "Code module DTO for updating"
    allOf:
      - $ref: "#/definitions/CodeModuleDTO"
    properties:
      tests:
        type: "array"
        items:
          $ref: "#/definitions/CodeModuleTestDTO"
      files:
        type: "array"
        items:
          $ref: "#/definitions/CodeModuleFileDTO"
  CodeModuleCreateDTO:
    description: "Code module DTO for creating"
    allOf:
      - $ref: "#/definitions/CodeModuleUpdateDTO"
    properties:
      moduleId:
        type: "integer"
        description: "Unique module identifier"
        example: 1
  CodeModuleTestDTO:
    type: "object"
    description: "Data transfer object for code module test detail / creating / updating"
    properties:
      id:
        type: "integer"
        description: "Unique identifier of test (unused for editing)"
        example: 1
      realId:
        type: "integer"
        description: "Same as id, used to determine whether to update (!= null) or create (= null) in edit requests"
        example: 1
      name:
        type: "string"
        description: "Test name"
        example: "Basic tests"
      parameter:
        type: "integer"
        description: "Parameter passed to tester"
        example: 101
      description:
        type: "string"
        description: "Test description"
        example: "Tests basic values"
      timeLimit:
        type: "integer"
        description: "Number of seconds after which code execution should be stopped (null = ignore)"
        example: 30
      checkMemory:
        type: "boolean"
        description: "Flag indicating whether freeing memory should be controlled"
        example: true
      shouldFail:
        type: "boolean"
        description: "Should the test fail (only write assert tests)"
        example: false
      hidden:
        type: "boolean"
        description: "Are results of the test hidden?"
        example: true
  CodeModuleFileDTO:
    type: "object"
    description: "Data transfer object for code module file detail / creating / updating"
    properties:
      id:
        type: "integer"
        description: "Unique identifier of file (unused for editing)"
        example: 1
      realId:
        type: "integer"
        description: "Same as id, used to determine whether to update (!= null) or create (= null) in edit requests"
        example: 1
      name:
        type: "string"
        description: "File name"
        example: "student.cpp"
      codeLimit:
        type: "integer"
        description: "Maximum code content length (in characters)"
        example: 1024
      content:
        type: "string"
        description: "File default content"
        example: "#include <stdio.h> int main () { return 0; }"
      reference:
        type: "string"
        description: "Reference solution"
        example: "#include <stdio.h> int main () { return 0; }"
      headerFile:
        type: "boolean"
        description: "Flag indicating whether file should be compiled, or just included on filesystem"
        example: true